Optimises for O(n), fixes overlapping matches.
Previously, for each index in the string of length N, the algorithm
would look ahead M characters (where M is the length of the
substring). This gave it an algorithmic complexity of O(nm), which could
be slow with longer substrings (for N = 3000, and M = 30, for example, a
total of 90000 lookups would be necessary).

The new implementation fuses lookup and iteration, such that we only
look at the character at the point, and no further. This works by
optimistically going forward in the string and comparing the current
character against the current position in the substring. If we fail, we
start over in the next position of the string. Otherwise, if we succeed
all the way, we keep looking from after that point.

This also fixes #1.